+++ pages/rmrk/create.vue
@@ -1,24 +1,20 @@
 <template>
-  <div class="columns">
-    <div class="column is-6 is-offset-3">
-      <section>
-        <br>
-        <b-tabs
-          v-model="activeTab"
-          destroy-on-hide
-          expanded
-        >
-          <b-tab-item
-            v-for="x in components"
-            :key="x"
-            :label="x"
-          >
-            <component :is="x" />
-          </b-tab-item>
-        </b-tabs>
-      </section>
-    </div>
-  </div>
+  <section>
+    <br>
+    <b-tabs
+      v-model="activeTab"
+      destroy-on-hide
+      expanded
+    >
+      <b-tab-item
+        v-for="x in components"
+        :key="x"
+        :label="x"
+      >
+        <component :is="x" />
+      </b-tab-item>
+    </b-tabs>
+  </section>
 </template>
 
 <script lang="ts" >
@@ -31,7 +27,11 @@ const components = { Collection, NFT }
 @Component({ components })
 
 export default class Remark extends Vue {
-    public activeTab = 0;
-    public components: string[] = ['Collection', 'NFT']
+  public activeTab = 0;
+  public components: string[] = ['Collection', 'NFT']
+
+  layout (context) {
+    return 'centered-half-layout'
+  }
 }
 </script>
